0 Tareas de clase/07. Pets_proyecto_final/requirements.txt
Django==4.2.7
mongoengine==0.27.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
pymongo==4.6.0
django-cors-headers==4.3.1
redis==5.0.1
requests==2.31.0